backend/requirements.txt
Flask==2.2.5
tensorflow==2.14.0
pillow
numpy
